[NFC][BACKEND] Move `LocalLoadOpConversion` to `MemoryOpToLLVM.cpp` (#4427)
It was separated from the previous PR because we felt it was too long,
and splitting it up would make things clearer
